pyproject.toml
[tool.poetry]
name = "hugging-face"
version = "0.1.0"
description = ""
authors = ["Daniel Leong <daniel.leong@program.com.tw>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.6"
uvicorn = "^0.32.1"
requests = "^2.32.3"
transformers = "^4.47.0"
python-dotenv = "^1.0.1"
line-bot-sdk = "^3.14.2"
replicate = "^1.0.4"
diffusers = "^0.31.0"
accelerate = "^1.2.0"
safetensors = "^0.4.5"
huggingface-hub = "^0.26.5"
pandas = "^2.2.3"
peft = "^0.14.0"
torch = {version = "^2.5.1+cu124", source = "pytorch-gpu"}
torchvision = {version = "^0.20.1+cu124", source = "pytorch-gpu"}
torchaudio = {version = "^2.5.1+cu124", source = "pytorch-gpu"}
sentencepiece = "^0.2.0"
aiofiles = "^24.1.0"
bitsandbytes = "^0.45.0"
numpy = "2.0"
ffmpeg-python = "^0.2.0"
ffprobe = "^0.5"
nltk = "^3.9.1"
gputil = "^1.4.0"
jsonschema = "^4.23.0"

[[tool.poetry.source]]
name = "pytorch-gpu"
url = "https://download.pytorch.org/whl/cu124"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
